feat(Ordering) move Order entity to Ordering namespace
