+++ src/Domain/Entities/Ordering/OrderStatus.cs
@@ -1,4 +1,4 @@
-﻿namespace Ticketing.Domain.Entities.Event;
+﻿namespace Ticketing.Domain.Entities.Ordering;
 
 public class OrderStatus
 {
