+++ src/Domain/Ticketing.Domain.csproj
@@ -6,4 +6,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="Entities\Ordering\" />
+  </ItemGroup>
+
 </Project>
